Fix merge fallout from origin/main: wire request_started through the new Responses completion path and extract settlement parsing to native_settlement.py to stay under the 999-line cap
